requirements.txt
Flask==3.0.3
requests==2.32.3
openai==1.41.0
python-dotenv==1.0.1
pandas==2.2.2
psutil==6.0.0
pulp==2.9.0